+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## NEXT
+
+- Fixed a bug in `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.from_adjacency_matrices` for
+  `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph` where the method was not handling properly
+  undirected edges in the adjacency matrices.
+
 ## 0.4.4
 
 - Relaxed `mypy` dependency to `"*"`.